last edit system

+++ app/Livewire/Permission/DeletePermission.php
@@ -24,6 +24,8 @@ class DeletePermission extends Component
 
     public function delete()
     {
+        $this->authorize('delete-permission');
+
         $this->deletePermission($this->id);
     }
